WpfToolkit ColorPicker and TopbarColor setting
