+++ src/Captura/Views/MoreOptionsWindow.xaml
@@ -1,13 +1,27 @@
 ﻿<Window x:Class="Captura.MoreOptionsWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:xctk="http://schemas.xceed.com/wpf/xaml/toolkit"
         xmlns:local="clr-namespace:Captura"
         Title="Experimental"
         DataContext="{StaticResource MainViewModel}"
         Height="600"
         Width="400">
+    <Window.Resources>
+        <Style TargetType="xctk:ColorPicker">
+            <Setter Property="ColorMode" Value="ColorCanvas"/>
+            <Setter Property="ShowAdvancedButton" Value="False"/>
+        </Style>
+    </Window.Resources>
     <Grid>
         <StackPanel Margin="10">
+            <DockPanel Margin="0,5">
+                <Label Content="Top Bar Color"
+                       ContentStringFormat="{}{0}: "
+                       Margin="0,0,5,0"/>
+                <xctk:ColorPicker SelectedColor="{Binding Settings.TopBarColor}"/>
+            </DockPanel>
+                        
             <Label Content="Mouse Clicks"
                    FontWeight="Bold"
                    Margin="0,3"
@@ -31,7 +45,7 @@
                 <Label Content="Color"
                        ContentStringFormat="{}{0}: "
                        Margin="0,0,5,0"/>
-                <TextBox Text="{Binding Settings.MouseClick_Color, UpdateSourceTrigger=PropertyChanged, Converter={StaticResource ColorToHexConverter}}"/>
+                <xctk:ColorPicker SelectedColor="{Binding Settings.MouseClick_Color, Converter={StaticResource WpfColorConverter}}"/>
             </DockPanel>
 
             <Label Content="Keystrokes"
@@ -43,7 +57,7 @@
                 <Label Content="Color"
                        ContentStringFormat="{}{0}: "
                        Margin="0,0,5,0"/>
-                <TextBox Text="{Binding Settings.Keystrokes_Color, UpdateSourceTrigger=PropertyChanged, Converter={StaticResource ColorToHexConverter}}" />
+                <xctk:ColorPicker SelectedColor="{Binding Settings.Keystrokes_Color, Converter={StaticResource WpfColorConverter}}"/>
             </DockPanel>
             <DockPanel Margin="0,5">
                 <Label Content="Font Size"
@@ -77,7 +91,7 @@
                 <Label Content="Back Color"
                        ContentStringFormat="{}{0}: "
                        Margin="0,0,5,0"/>
-                <TextBox Text="{Binding Settings.KeystrokesRect_Color, UpdateSourceTrigger=PropertyChanged, Converter={StaticResource ColorToHexConverter}}"/>
+                <xctk:ColorPicker SelectedColor="{Binding Settings.KeystrokesRect_Color, Converter={StaticResource WpfColorConverter}}"/>
             </DockPanel>
 
             <DockPanel Margin="0,5">
