+++ src/Captura/packages.config
@@ -1,6 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="CommandLineParser" version="1.9.71" targetFramework="net461" />
+  <package id="Extended.Wpf.Toolkit" version="3.0" targetFramework="net461" />
   <package id="Hardcodet.NotifyIcon.Wpf" version="1.0.8" targetFramework="net461" />
   <package id="ModernUI.WPF" version="1.0.9" targetFramework="net461" />
   <package id="WebEye.Controls.Wpf.WebCameraControl" version="1.0.1" targetFramework="net461" />
